pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lore-mcp-server"
version = "1.0.0"
description = "Production-minded governed memory MCP server for AI agents with deterministic deletion and auditable lineage"
readme = "README.md"
license = "Apache-2.0"
authors = [{ name = "Lore Maintainers" }]
requires-python = ">=3.12"
keywords = ["mcp", "memory", "agent-memory", "compliance", "lineage", "gdpr", "ai-agent", "context-protocol"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "mcp==1.26.0",
  "cryptography>=46.0.6,<47",
  "argon2-cffi==25.1.0",
  "PyJWT[crypto]==2.12.0",
  "httpx==0.28.1",
]

[project.scripts]
orenyl-server = "orenyl.server:main"

[project.urls]
Homepage = "https://github.com/edison-zhou-nc/Lore"
Repository = "https://github.com/edison-zhou-nc/Lore"
Changelog = "https://github.com/edison-zhou-nc/Lore/blob/main/CHANGELOG.md"
Issues = "https://github.com/edison-zhou-nc/Lore/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
orenyl = ["schema.sql", "py.typed"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S105", "S603"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [".git", ".venv", ".tmp_pytest_local", "tests/.tmp", "tests/.pytest_tmp", "tests/pytest_tmp_*"]
markers = ["perf: deterministic performance smoke gates"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/orenyl"]
omit = ["**/tests/**", "**/schema.sql"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
